Reorganise some of the codebase into a submodule
